well this isn't where it should be
I finally got it to save pictures again and display them. Need to test to see if the file is actually copied over to internal storage and then delete the original so it doesn't appear in the damn gallery.

why can the Camera not write directly into the app's internal storage. I AM THE ONE WHO CALLED YOU, YOU HAVE MY PERMISSION. ugh