Fixed saving images, sort of. Also AsyncTask
Found a good guide on saving pictures to internal storage, but it turns out HTC's camera doesn't respect that and saves a local copy regardless. The Google camera works just fine.

Moved saving the picture off to an AsyncTask since it was slowing down the app every time it tried to save things.